#1328 Coordinator decompose Phase 2 — split reconcile() + refresh_bindings()
Closes #1328

Pure code-motion refactor following the #1189 Phase 1 mod-dir
pattern. Decomposes the two longest non-hot functions in
userspace-dp/src/afxdp/coordinator/mod.rs:

  - reconcile() (506 LOC, off the packet hot path; fires on
    snapshot installs + HA role changes) — split into a 65-line
    orchestrator + 4 phase sub-files under
    coordinator/reconcile/{teardown,reset,snapshot,bringup}.rs.

  - refresh_bindings() (326 LOC, ~1Hz status poll) — split into
    sibling coordinator/refresh_bindings.rs with
    copy_live_snapshot + zero_unbound_slot helpers and a thin
    dispatcher.

Layout:

  coordinator/
    mod.rs               (2026 LOC -> 1194 LOC, -832)
    refresh_bindings.rs  (368 LOC, new)
    reconcile/
      mod.rs             (96 LOC: orchestrator +
                          PreservedReconcileState +
                          ReconcileSnapshotFds)
      teardown.rs        (40 LOC: preserve sessions + healthy
                          slow path, stop_inner, 500ms mlx5
                          quiesce when workers were live)
      reset.rs           (63 LOC: reset_binding_counters)
      snapshot.rs        (172 LOC: apply_snapshot — validation +
                          forwarding install, slow-path re-arm,
                          map FD opens; returns
                          Option<ReconcileSnapshotFds>; on
                          failure sets last_reconcile_stage +
                          per-registered-binding last_error)
      bringup.rs         (333 LOC: bring_up_workers — worker
                          plan loop, replay synced sessions,
                          per-worker spawn (#925 panic-slot
                          insert + remove kept INLINE, paired
                          with one-line comment), start
                          neighbor monitor, spawn local tunnel
                          sources)

Hidden invariants preserved:

- Side-effect ordering inside reconcile() verbatim: teardown ->
  reset -> snapshot apply -> bringup -> refresh_bindings tail.
- last_reconcile_stage write sequence verbatim including the
  indirect "stopped" write inside stop_inner (L281) and the
  pre-existing "spawn_worker_failed" -> "spawned" overwrite
  quirk on the spawn-Err arm.
- #925 panic-slot insert+remove pairing kept inline in
  bringup.rs (spawn closure captures ~30 worker-scoped Arcs;
  a wrapper helper would hide more than it clarifies).
- Allocation discipline: no additional allocations beyond the
  pre-existing stop_inner-driven Arc::new(BTreeMap::new())
  stores. The refresh_bindings histogram copies use
  &snap.tx_*_latency_hist + copy_from_slice into the existing
  Vec<u64> backing storage on BindingStatus, so no per-poll
  allocation.
- BindingLiveSnapshot is taken by VALUE in copy_live_snapshot
  (it owns String fields — borrowing would force per-poll
  clones).
- PreservedReconcileState has exactly 2 fields
  (synced_sessions + slow_path). had_live_workers is local to
  teardown.

Visibility promotions to wire sub-modules:

- pub(super) fns: build_cos_owner_worker_by_queue,
  build_mirror_target_map,
  build_cos_active_shards_by_egress_ifindex_with_fallback_ifindexes.
- pub(super) methods: spawn_local_tunnel_sources,
  refresh_cos_owner_worker_map_from_binding_statuses,
  refresh_cos_runtime_maps.

Plan + 4-round adversarial review:

- docs/pr/1328-coordinator-reconcile-split/plan.md v2.2
- Codex r1 PLAN-NEEDS-MAJOR (8 findings) -> v2 -> Codex r2
  PLAN-NEEDS-MAJOR (4 stale-text findings) -> v2.1 -> Codex
  r3 PLAN-NEEDS-MINOR (5 findings: sub-file naming,
  mod-count, test-claim, LOC over/under, histogram wording)
  -> v2.2 -> Codex r4 PLAN-READY (best effort; sandbox infra
  intermittently broken).
- AGY r1 PLAN-NEEDS-MINOR (layout) -> v2 -> AGY r2
  PLAN-NEEDS-MINOR (4 stale-text) -> v2.1 -> AGY r3
  PLAN-READY -> v2.2 -> AGY r4 PLAN-READY.

Validation:

- cargo build --release: clean.
- cargo test --release --bins: 1418 + 46 + 8 + 16 = 1488 pass
  (one new test: reconcile_with_none_snapshot_preserves_stage_
  sequence).
- coordinator filter 5x: 33 tests pass each run (no flake).
- go test ./...: 32 packages pass.
- One pre-existing upstream failure: snat_contract_doc_guard
  asserts "fail-closed" in docs/userspace-dataplane-gaps.md
  which currently reads "fail closed" (no hyphen). AGY r4
  flagged this as upstream master drift, not part of #1328.

Smoke deferred to batch per the standing AWAITING-BATCH-MERGE
rule on retirement/refactor chains.
