+++ userspace-dp/src/afxdp/coordinator/mod.rs
@@ -4,6 +4,8 @@ mod cos_state;
 mod ha_state;
 mod inject;
 mod neighbor_manager;
+mod reconcile;
+mod refresh_bindings;
 mod session_manager;
 mod status;
 mod supervisor;
@@ -13,7 +15,7 @@ pub(crate) use cos_state::SharedCoSState;
 pub(in crate::afxdp) use ha_state::HaState;
 pub(crate) use neighbor_manager::NeighborManager;
 pub(in crate::afxdp) use session_manager::SessionManager;
-use supervisor::{spawn_supervised_aux, spawn_supervised_worker};
+use supervisor::spawn_supervised_aux;
 pub(in crate::afxdp) use worker_manager::WorkerManager;
 
 pub struct Coordinator {
@@ -311,515 +313,8 @@ impl Coordinator {
         entries.len()
     }
 
-    pub fn reconcile(
-        &mut self,
-        snapshot: Option<&ConfigSnapshot>,
-        bindings: &mut [BindingStatus],
-        ring_entries: usize,
-    ) {
-        self.reconcile_calls += 1;
-        self.last_reconcile_stage = "start".to_string();
-        let had_live_workers = !self.workers.handles.is_empty();
-        let preserved_synced_sessions = self.snapshot_shared_session_entries();
-        // Keep a healthy slow-path worker across back-to-back reconciles. The
-        // userspace helper can receive multiple snapshot refreshes during HA
-        // role changes; recreating the fixed-name TUN on every reconcile can
-        // race with teardown and leave the new owner without xpf-usp0.
-        let preserved_slow_path = self.slow_path.as_ref().and_then(|slow| {
-            if slow.status().active {
-                Some(slow.clone())
-            } else {
-                None
-            }
-        });
-        self.stop_inner(false);
-        if had_live_workers {
-            // Zero-copy queue teardown is not synchronously reusable on mlx5.
-            // A short quiesce avoids EBUSY when a later snapshot refresh
-            // rebuilds the same queue set immediately after shutdown.
-            thread::sleep(Duration::from_millis(500));
-        }
-        for binding in bindings.iter_mut() {
-            binding.bound = false;
-            binding.xsk_registered = false;
-            binding.socket_fd = 0;
-            binding.rx_packets = 0;
-            binding.rx_bytes = 0;
-            binding.rx_batches = 0;
-            binding.rx_wakeups = 0;
-            binding.metadata_packets = 0;
-            binding.metadata_errors = 0;
-            binding.validated_packets = 0;
-            binding.validated_bytes = 0;
-            binding.local_delivery_packets = 0;
-            binding.forward_candidate_packets = 0;
-            binding.route_miss_packets = 0;
-            binding.neighbor_miss_packets = 0;
-            binding.discard_route_packets = 0;
-            binding.next_table_packets = 0;
-            binding.exception_packets = 0;
-            binding.config_gen_mismatches = 0;
-            binding.fib_gen_mismatches = 0;
-            binding.unsupported_packets = 0;
-            binding.flow_cache_hits = 0;
-            binding.flow_cache_misses = 0;
-            binding.flow_cache_evictions = 0;
-            binding.flow_cache_collision_evictions = 0;
-            binding.v_min_throttle_hard_cap_overrides = 0;
-            binding.v_min_throttles = 0;
-            binding.session_hits = 0;
-            binding.session_misses = 0;
-            binding.session_creates = 0;
-            binding.session_expires = 0;
-            binding.session_delta_pending = 0;
-            binding.session_delta_generated = 0;
-            binding.session_delta_dropped = 0;
-            binding.session_delta_drained = 0;
-            binding.policy_denied_packets = 0;
-            binding.snat_packets = 0;
-            binding.dnat_packets = 0;
-            binding.slow_path_packets = 0;
-            binding.slow_path_bytes = 0;
-            binding.slow_path_local_delivery_packets = 0;
-            binding.slow_path_missing_neighbor_packets = 0;
-            binding.slow_path_no_route_packets = 0;
-            binding.slow_path_next_table_packets = 0;
-            binding.slow_path_forward_build_packets = 0;
-            binding.slow_path_drops = 0;
-            binding.slow_path_rate_limited = 0;
-            binding.kernel_rx_dropped = 0;
-            binding.kernel_rx_invalid_descs = 0;
-            binding.last_error.clear();
-            binding.ready = false;
-        }
-        let Some(snapshot) = snapshot else {
-            self.policy_counters.reconcile_rules(&[]);
-            self.last_reconcile_stage = "no_snapshot".to_string();
-            return;
-        };
-        self.validation = ValidationState {
-            snapshot_installed: true,
-            config_generation: snapshot.generation,
-            fib_generation: snapshot.fib_generation,
-        };
-        self.policy_counters.reconcile_rules(&snapshot.policies);
-        self.forwarding = build_forwarding_state_with_policy_counters_and_previous(
-            snapshot,
-            &self.policy_counters,
-            Some(&self.forwarding),
-        );
-        self.shared_validation.store(Arc::new(self.validation));
-        self.ha.forwarding.store(Arc::new(self.forwarding.clone()));
-        self.slow_path = if let Some(slow_path) = preserved_slow_path {
-            self.last_slow_path_status = slow_path.status();
-            Some(slow_path)
-        } else {
-            match SlowPathReinjector::new(DEFAULT_SLOW_PATH_TUN) {
-                Ok(reinjector) => {
-                    self.last_slow_path_status = reinjector.status();
-                    Some(Arc::new(reinjector))
-                }
-                Err(err) => {
-                    self.last_slow_path_status = SlowPathStatus {
-                        last_error: err,
-                        ..SlowPathStatus::default()
-                    };
-                    None
-                }
-            }
-        };
-        self.local_tunnel_deliveries
-            .store(Arc::new(BTreeMap::new()));
-        self.ha
-            .fabrics
-            .store(Arc::new(self.forwarding.fabrics.clone()));
-        if snapshot.map_pins.xsk.is_empty() {
-            self.last_reconcile_stage = "missing_xsk_pin".to_string();
-            for binding in bindings.iter_mut() {
-                if binding.registered {
-                    binding.last_error = "missing XSK map pin path".to_string();
-                }
-            }
-            return;
-        }
-        if snapshot.map_pins.heartbeat.is_empty() {
-            self.last_reconcile_stage = "missing_heartbeat_pin".to_string();
-            for binding in bindings.iter_mut() {
-                if binding.registered {
-                    binding.last_error = "missing heartbeat map pin path".to_string();
-                }
-            }
-            return;
-        }
-        if snapshot.map_pins.sessions.is_empty() {
-            self.last_reconcile_stage = "missing_session_pin".to_string();
-            for binding in bindings.iter_mut() {
-                if binding.registered {
-                    binding.last_error = "missing session map pin path".to_string();
-                }
-            }
-            return;
-        }
-        let map_fd = match OwnedFd::open_bpf_map(&snapshot.map_pins.xsk) {
-            Ok(fd) => fd,
-            Err(err) => {
-                self.last_reconcile_stage = format!("open_xsk_map_failed:{err}");
-                for binding in bindings.iter_mut() {
-                    if binding.registered {
-                        binding.last_error = format!("open XSK map: {err}");
-                    }
-                }
-                return;
-            }
-        };
-        let heartbeat_map_fd = match OwnedFd::open_bpf_map(&snapshot.map_pins.heartbeat) {
-            Ok(fd) => fd,
-            Err(err) => {
-                self.last_reconcile_stage = format!("open_heartbeat_map_failed:{err}");
-                for binding in bindings.iter_mut() {
-                    if binding.registered {
-                        binding.last_error = format!("open heartbeat map: {err}");
-                    }
-                }
-                return;
-            }
-        };
-        let session_map_fd = match OwnedFd::open_bpf_map(&snapshot.map_pins.sessions) {
-            Ok(fd) => fd,
-            Err(err) => {
-                self.last_reconcile_stage = format!("open_session_map_failed:{err}");
-                for binding in bindings.iter_mut() {
-                    if binding.registered {
-                        binding.last_error = format!("open session map: {err}");
-                    }
-                }
-                return;
-            }
-        };
-        // Open BPF conntrack maps (sessions, sessions_v6) so the helper can
-        // publish session entries that "show security flow session" reads.
-        // Non-fatal: if the maps don't exist, session display will lack zone/interface info.
-        let conntrack_v4_fd = if !snapshot.map_pins.conntrack_v4.is_empty() {
-            OwnedFd::open_bpf_map(&snapshot.map_pins.conntrack_v4).ok()
-        } else {
-            None
-        };
-        let conntrack_v6_fd = if !snapshot.map_pins.conntrack_v6.is_empty() {
-            OwnedFd::open_bpf_map(&snapshot.map_pins.conntrack_v6).ok()
-        } else {
-            None
-        };
-        // Open dnat_table BPF map for embedded ICMP NAT reversal support.
-        // Non-fatal: if the map doesn't exist, embedded ICMP won't work
-        // but normal forwarding is unaffected.
-        let dnat_table_fd = if !snapshot.map_pins.dnat_table.is_empty() {
-            OwnedFd::open_bpf_map(&snapshot.map_pins.dnat_table).ok()
-        } else {
-            None
-        };
-        let dnat_table_v6_fd = if !snapshot.map_pins.dnat_table_v6.is_empty() {
-            OwnedFd::open_bpf_map(&snapshot.map_pins.dnat_table_v6).ok()
-        } else {
-            None
-        };
-        let dnat_fds = DnatTableFds {
-            v4: dnat_table_fd.as_ref().map(|f| f.fd),
-            v6: dnat_table_v6_fd.as_ref().map(|f| f.fd),
-        };
-        let ring_entries = ring_entries.max(64).min(u32::MAX as usize) as u32;
-        let mut workers: BTreeMap<u32, Vec<BindingPlan>> = BTreeMap::new();
-        for binding in bindings.iter_mut() {
-            if !binding.registered || binding.ifindex <= 0 {
-                binding.ready = false;
-                continue;
-            }
-            let live = Arc::new(BindingLiveState::new());
-            self.workers.live.insert(binding.slot, live.clone());
-            let identity = BindingIdentity {
-                slot: binding.slot,
-                queue_id: binding.queue_id,
-                worker_id: binding.worker_id,
-                interface: Arc::<str>::from(binding.interface.as_str()),
-                ifindex: binding.ifindex,
-            };
-            self.workers.identities.insert(binding.slot, identity);
-            workers
-                .entry(binding.worker_id)
-                .or_default()
-                .push(BindingPlan {
-                    status: binding.clone(),
-                    live,
-                    xsk_map_fd: map_fd.fd,
-                    heartbeat_map_fd: heartbeat_map_fd.fd,
-                    session_map_fd: session_map_fd.fd,
-                    conntrack_v4_fd: conntrack_v4_fd.as_ref().map(|f| f.fd).unwrap_or(-1),
-                    conntrack_v6_fd: conntrack_v6_fd.as_ref().map(|f| f.fd).unwrap_or(-1),
-                    ring_entries,
-                    bind_strategy: preferred_bind_strategy(binding),
-                    poll_mode: self.poll_mode,
-                    shared_umem: SharedUmemBindingPlan::private(),
-                });
-        }
-        for plans in workers.values_mut() {
-            plans.sort_by_key(|plan| (plan.status.queue_id, plan.status.ifindex, plan.status.slot));
-        }
-        apply_shared_umem_policy_to_workers(snapshot, &mut workers);
-        let shared_umem_status_by_slot = workers
-            .values()
-            .flat_map(|plans| plans.iter())
-            .map(|plan| {
-                (
-                    plan.status.slot,
-                    (
-                        plan.status.shared_umem_mode.clone(),
-                        plan.status.shared_umem_group.clone(),
-                        plan.status.shared_umem_socket_role.clone(),
-                        plan.status.shared_umem_disabled_reason.clone(),
-                    ),
-                )
-            })
-            .collect::<BTreeMap<_, _>>();
-        for binding in bindings.iter_mut() {
-            if let Some((mode, group, role, reason)) = shared_umem_status_by_slot.get(&binding.slot)
-            {
-                binding.shared_umem_mode.clone_from(mode);
-                binding.shared_umem_group.clone_from(group);
-                binding.shared_umem_socket_role.clone_from(role);
-                binding.shared_umem_disabled_reason.clone_from(reason);
-            }
-        }
-        let planned_bindings: usize = workers.values().map(|group| group.len()).sum();
-        self.workers.last_planned_workers = workers.len();
-        self.workers.last_planned_bindings = planned_bindings;
-        self.last_reconcile_stage = format!(
-            "planned:workers={}:bindings={}:live={}",
-            self.workers.last_planned_workers(),
-            self.workers.last_planned_bindings(),
-            self.workers.live.len()
-        );
-        eprintln!(
-            "xpf-userspace-dp: reconcile planned_workers={} planned_bindings={} live_slots={}",
-            workers.len(),
-            planned_bindings,
-            self.workers.live.len()
-        );
-        let session_map_raw_fd = session_map_fd.fd;
-        self.bpf_maps.map_fd = Some(map_fd);
-        self.bpf_maps.heartbeat_map_fd = Some(heartbeat_map_fd);
-        self.bpf_maps.session_map_fd = Some(session_map_fd);
-        self.bpf_maps.conntrack_v4_fd = conntrack_v4_fd;
-        self.bpf_maps.conntrack_v6_fd = conntrack_v6_fd;
-        self.bpf_maps.dnat_table_fd = dnat_table_fd;
-        self.bpf_maps.dnat_table_v6_fd = dnat_table_v6_fd;
-        let worker_binding_ifindexes = workers
-            .iter()
-            .map(|(worker_id, binding_plans)| {
-                (
-                    *worker_id,
-                    binding_plans
-                        .iter()
-                        .map(|plan| plan.status.ifindex)
-                        .collect::<std::collections::BTreeSet<_>>(),
-                )
-            })
-            .collect::<BTreeMap<_, _>>();
-        let owner_map = build_cos_owner_worker_by_queue(&self.forwarding, &workers);
-        self.mirror_targets.store(Arc::new(build_mirror_target_map(
-            &self.workers.identities,
-            &self.workers.live,
-        )));
-        let active_shards_by_egress_ifindex =
-            build_cos_active_shards_by_egress_ifindex_with_fallback_ifindexes(
-                &self.forwarding,
-                &worker_binding_ifindexes,
-                &worker_binding_ifindexes,
-            );
-        self.refresh_cos_runtime_maps(owner_map, active_shards_by_egress_ifindex);
-        let worker_command_queues: Arc<BTreeMap<u32, Arc<Mutex<VecDeque<WorkerCommand>>>>> =
-            Arc::new(
-                workers
-                    .keys()
-                    .copied()
-                    .map(|worker_id| (worker_id, Arc::new(Mutex::new(VecDeque::new()))))
-                    .collect(),
-            );
-        let replayed_synced_sessions = self.replay_synced_sessions(
-            &preserved_synced_sessions,
-            worker_command_queues.as_ref(),
-            session_map_raw_fd,
-        );
-        if replayed_synced_sessions > 0 {
-            self.last_reconcile_stage = format!(
-                "replayed_synced:{}:workers={}",
-                replayed_synced_sessions,
-                worker_command_queues.len()
-            );
-        }
-        for (worker_id, binding_plans) in workers {
-            let plan_count = binding_plans.len();
-            let stop = Arc::new(AtomicBool::new(false));
-            let heartbeat = Arc::new(AtomicU64::new(monotonic_nanos()));
-            let session_export_ack = Arc::new(AtomicU64::new(0));
-            let cos_status = Arc::new(ArcSwap::from_pointee(Vec::new()));
-            let commands = worker_command_queues
-                .get(&worker_id)
-                .cloned()
-                .unwrap_or_else(|| Arc::new(Mutex::new(VecDeque::new())));
-            let recent_exceptions = self.recent_exceptions.clone();
-            let recent_session_deltas = self.recent_session_deltas.clone();
-            let last_resolution = self.last_resolution.clone();
-            let slow_path = self.slow_path.clone();
-            let local_tunnel_deliveries = self.local_tunnel_deliveries.clone();
-            let shared_forwarding = self.ha.forwarding.clone();
-            let shared_validation = self.shared_validation.clone();
-            let shared_sessions = self.sessions.synced.clone();
-            let shared_nat_sessions = self.sessions.nat.clone();
-            let shared_forward_wire_sessions = self.sessions.forward_wire.clone();
-            let shared_owner_rg_indexes = self.sessions.owner_rg_indexes.clone();
-            let stop_clone = stop.clone();
-            let heartbeat_clone = heartbeat.clone();
-            let session_export_ack_clone = session_export_ack.clone();
-            let commands_clone = commands.clone();
-            let peer_commands_clone = worker_command_queues
-                .iter()
-                .filter(|(id, _)| **id != worker_id)
-                .map(|(_, queue)| queue.clone())
-                .collect::<Vec<_>>();
-            let worker_commands_by_id = worker_command_queues.clone();
-            let ha_state = self.ha.rg_runtime.clone();
-            let dynamic_neighbors = self.neighbors.dynamic.clone();
-            let worker_poll_mode = self.poll_mode;
-            let shared_fabrics = self.ha.fabrics.clone();
-            let rg_epochs = self.rg_epochs.clone();
-            let event_stream_handle = self.event_stream_worker_handle();
-            let cos_status_clone = cos_status.clone();
-            let shared_cos_owner_worker_by_queue = self.cos.owner_worker_by_queue.clone();
-            let shared_cos_owner_live_by_queue = self.cos.owner_live_by_queue.clone();
-            let shared_cos_root_leases = self.cos.root_leases.clone();
-            let shared_cos_exact_backlogs = self.cos.exact_backlogs.clone();
-            let shared_cos_queue_leases = self.cos.queue_leases.clone();
-            let shared_cos_queue_vtime_floors = self.cos.queue_vtime_floors.clone();
-            let shared_mirror_targets = self.mirror_targets.clone();
-            let runtime_atomics =
-                std::sync::Arc::new(super::worker_runtime::WorkerRuntimeAtomics::new());
-            let runtime_atomics_clone = runtime_atomics.clone();
-            // #925 Phase 1: per-worker panic slot, keyed by worker_id.
-            let panic_slot = Arc::new(Mutex::new(None::<String>));
-            self.worker_panics.insert(worker_id, panic_slot.clone());
-            let join = spawn_supervised_worker(
-                worker_id,
-                runtime_atomics.clone(),
-                panic_slot,
-                move || {
-                    worker_loop(
-                        worker_id,
-                        binding_plans,
-                        shared_validation,
-                        shared_forwarding,
-                        ha_state,
-                        dynamic_neighbors,
-                        shared_sessions,
-                        shared_nat_sessions,
-                        shared_forward_wire_sessions,
-                        shared_owner_rg_indexes,
-                        slow_path,
-                        local_tunnel_deliveries,
-                        recent_exceptions,
-                        recent_session_deltas,
-                        last_resolution,
-                        commands_clone,
-                        peer_commands_clone,
-                        worker_commands_by_id,
-                        stop_clone,
-                        heartbeat_clone,
-                        session_export_ack_clone,
-                        worker_poll_mode,
-                        dnat_fds,
-                        shared_fabrics,
-                        event_stream_handle,
-                        rg_epochs,
-                        shared_cos_owner_worker_by_queue,
-                        shared_cos_owner_live_by_queue,
-                        shared_cos_root_leases,
-                        shared_cos_exact_backlogs,
-                        shared_cos_queue_leases,
-                        shared_cos_queue_vtime_floors,
-                        shared_mirror_targets,
-                        cos_status_clone,
-                        runtime_atomics_clone,
-                    );
-                },
-            );
-            match join {
-                Ok(join) => {
-                    eprintln!(
-                        "xpf-userspace-dp: started worker thread worker_id={} planned_bindings={}",
-                        worker_id, plan_count
-                    );
-                    self.workers.handles.insert(
-                        worker_id,
-                        WorkerHandle {
-                            stop,
-                            heartbeat,
-                            commands,
-                            session_export_ack,
-                            cos_status,
-                            runtime_atomics,
-                            join: Some(join),
-                        },
-                    );
-                }
-                Err(err) => {
-                    eprintln!(
-                        "xpf-userspace-dp: failed to start worker thread worker_id={} err={}",
-                        worker_id, err
-                    );
-                    self.last_reconcile_stage = format!("spawn_worker_failed:{worker_id}:{err}");
-                    // #925 Phase 1: the panic slot was inserted before
-                    // spawn; drop it now so a snapshot reader doesn't
-                    // see a phantom slot for a worker that never ran.
-                    self.worker_panics.remove(&worker_id);
-                    if let Ok(mut recent) = self.recent_exceptions.lock() {
-                        push_recent_exception(
-                            &mut recent,
-                            ExceptionStatus {
-                                timestamp: Utc::now(),
-                                reason: format!("spawn_worker_failed:{worker_id}:{err}"),
-                                ..ExceptionStatus::default()
-                            },
-                        );
-                    }
-                }
-            }
-        }
-        self.last_reconcile_stage = format!(
-            "spawned:workers={}:identities={}:live={}",
-            self.workers.handles.len(),
-            self.workers.identities.len(),
-            self.workers.live.len()
-        );
-        // Start the helper-owned neighbor sync path. It does an initial
-        // RTM_GETNEIGH dump so startup sees the existing kernel table, then
-        // subscribes to RTM_{NEW,DEL}NEIGH for incremental updates.
-        if self.neighbors.monitor_stop.is_none() {
-            let stop = Arc::new(AtomicBool::new(false));
-            let stop_clone = stop.clone();
-            let dynamic_neighbors = self.neighbors.dynamic.clone();
-            let neighbor_generation = self.neighbors.generation.clone();
-            // #925-A: wrap aux thread in catch_unwind so a panic in the
-            // netlink path doesn't kill the daemon. No respawn — see
-            // spawn_supervised_aux doc for operator-visible degradation.
-            spawn_supervised_aux("neigh-monitor", move || {
-                neigh_monitor_thread(stop_clone, dynamic_neighbors, neighbor_generation)
-            })
-            .ok();
-            self.neighbors.monitor_stop = Some(stop);
-        }
-        self.spawn_local_tunnel_sources();
-        self.refresh_bindings(bindings);
-    }
 
-    fn spawn_local_tunnel_sources(&mut self) {
+    pub(super) fn spawn_local_tunnel_sources(&mut self) {
         let mut local_tunnel_deliveries = BTreeMap::new();
         for endpoint in self.forwarding.tunnel_endpoints.values() {
             if endpoint.mode != "gre" && endpoint.mode != "ip6gre" {
@@ -1033,7 +528,7 @@ impl Coordinator {
         self.refresh_cos_runtime_maps(owner_map, active_shards_by_egress_ifindex);
     }
 
-    fn refresh_cos_owner_worker_map_from_binding_statuses(&mut self, bindings: &[BindingStatus]) {
+    pub(super) fn refresh_cos_owner_worker_map_from_binding_statuses(&mut self, bindings: &[BindingStatus]) {
         let ready_worker_binding_ifindexes = bindings.iter().filter(|binding| binding.ready).fold(
             BTreeMap::<u32, std::collections::BTreeSet<i32>>::new(),
             |mut out, binding| {
@@ -1059,7 +554,7 @@ impl Coordinator {
         self.refresh_cos_runtime_maps(owner_map, active_shards_by_egress_ifindex);
     }
 
-    fn refresh_cos_runtime_maps(
+    pub(super) fn refresh_cos_runtime_maps(
         &mut self,
         owner_map: BTreeMap<(i32, u8), u32>,
         active_shards_by_egress_ifindex: BTreeMap<i32, usize>,
@@ -1153,333 +648,6 @@ impl Coordinator {
         self.shared_validation.store(Arc::new(self.validation));
     }
 
-    pub fn refresh_bindings(&mut self, bindings: &mut [BindingStatus]) {
-        for binding in bindings.iter_mut() {
-            if let Some(live) = self.workers.live.get(&binding.slot) {
-                let snap = live.snapshot();
-                if snap.bound && !binding.bound {
-                    eprintln!(
-                        "refresh_bindings: slot={} transitioning bound=false->true fd={}",
-                        binding.slot, snap.socket_fd
-                    );
-                }
-                binding.bound = snap.bound;
-                binding.xsk_registered = snap.xsk_registered;
-                binding.xsk_bind_mode = snap.xsk_bind_mode;
-                binding.zero_copy = snap.zero_copy;
-                binding.socket_fd = snap.socket_fd;
-                binding.socket_ifindex = snap.socket_ifindex;
-                binding.socket_queue_id = snap.socket_queue_id;
-                binding.socket_bind_flags = snap.socket_bind_flags;
-                binding.shared_umem_mode = snap.shared_umem_mode;
-                binding.shared_umem_group = snap.shared_umem_group;
-                binding.shared_umem_socket_role = snap.shared_umem_socket_role;
-                binding.shared_umem_disabled_reason = snap.shared_umem_disabled_reason;
-                binding.rx_packets = snap.rx_packets;
-                binding.rx_bytes = snap.rx_bytes;
-                binding.rx_batches = snap.rx_batches;
-                binding.rx_wakeups = snap.rx_wakeups;
-                binding.metadata_packets = snap.metadata_packets;
-                binding.metadata_errors = snap.metadata_errors;
-                binding.validated_packets = snap.validated_packets;
-                binding.validated_bytes = snap.validated_bytes;
-                binding.local_delivery_packets = snap.local_delivery_packets;
-                binding.forward_candidate_packets = snap.forward_candidate_packets;
-                binding.route_miss_packets = snap.route_miss_packets;
-                binding.neighbor_miss_packets = snap.neighbor_miss_packets;
-                binding.discard_route_packets = snap.discard_route_packets;
-                binding.next_table_packets = snap.next_table_packets;
-                binding.exception_packets = snap.exception_packets;
-                binding.config_gen_mismatches = snap.config_gen_mismatches;
-                binding.fib_gen_mismatches = snap.fib_gen_mismatches;
-                binding.unsupported_packets = snap.unsupported_packets;
-                binding.flow_cache_hits = snap.flow_cache_hits;
-                binding.flow_cache_misses = snap.flow_cache_misses;
-                binding.flow_cache_evictions = snap.flow_cache_evictions;
-                binding.flow_cache_collision_evictions = snap.flow_cache_collision_evictions;
-                // #1219: bridge active_flow_count from BindingLiveSnapshot
-                // into BindingStatus so it reaches the wire-visible status.
-                binding.active_flow_count = snap.active_flow_count;
-                binding.flow_cache_capacity = snap.flow_cache_capacity;
-                // #941 Work item D / #943: bridge V_min counters from
-                // BindingLiveSnapshot through to BindingStatus so the
-                // wire surface (BindingCountersSnapshot) sees them.
-                binding.v_min_throttle_hard_cap_overrides = snap.v_min_throttle_hard_cap_overrides;
-                binding.v_min_throttles = snap.v_min_throttles;
-                binding.session_hits = snap.session_hits;
-                binding.session_misses = snap.session_misses;
-                binding.session_creates = snap.session_creates;
-                binding.session_expires = snap.session_expires;
-                binding.session_delta_pending = snap.session_delta_pending;
-                binding.session_delta_generated = snap.session_delta_generated;
-                binding.session_delta_dropped = snap.session_delta_dropped;
-                binding.session_delta_drained = snap.session_delta_drained;
-                binding.policy_denied_packets = snap.policy_denied_packets;
-                binding.screen_drops = snap.screen_drops;
-                binding.syn_cookie_challenges = snap.syn_cookie_challenges;
-                binding.syn_cookie_secret_unavailable = snap.syn_cookie_secret_unavailable;
-                binding.syn_cookie_syn_ack_sent = snap.syn_cookie_syn_ack_sent;
-                binding.syn_cookie_ack_rst_sent = snap.syn_cookie_ack_rst_sent;
-                binding.syn_cookie_reply_budget_drops = snap.syn_cookie_reply_budget_drops;
-                binding.syn_cookie_ack_valid = snap.syn_cookie_ack_valid;
-                binding.syn_cookie_ack_invalid = snap.syn_cookie_ack_invalid;
-                binding.syn_cookie_bypass = snap.syn_cookie_bypass;
-                binding.snat_packets = snap.snat_packets;
-                binding.dnat_packets = snap.dnat_packets;
-                binding.slow_path_packets = snap.slow_path_packets;
-                binding.slow_path_bytes = snap.slow_path_bytes;
-                binding.slow_path_local_delivery_packets = snap.slow_path_local_delivery_packets;
-                binding.slow_path_missing_neighbor_packets =
-                    snap.slow_path_missing_neighbor_packets;
-                binding.slow_path_no_route_packets = snap.slow_path_no_route_packets;
-                binding.slow_path_next_table_packets = snap.slow_path_next_table_packets;
-                binding.slow_path_forward_build_packets = snap.slow_path_forward_build_packets;
-                binding.slow_path_drops = snap.slow_path_drops;
-                binding.slow_path_rate_limited = snap.slow_path_rate_limited;
-                binding.kernel_rx_dropped = snap.kernel_rx_dropped;
-                binding.kernel_rx_invalid_descs = snap.kernel_rx_invalid_descs;
-                binding.tx_packets = snap.tx_packets;
-                binding.tx_bytes = snap.tx_bytes;
-                binding.tx_completions = snap.tx_completions;
-                binding.tx_errors = snap.tx_errors;
-                binding.tx_shared_recycle_unknown_slot_drops =
-                    snap.tx_shared_recycle_unknown_slot_drops;
-                binding.redirect_inbox_overflow_drops = snap.redirect_inbox_overflow_drops;
-                binding.pending_tx_local_overflow_drops = snap.pending_tx_local_overflow_drops;
-                binding.tx_submit_error_drops = snap.tx_submit_error_drops;
-                binding.mirrored_packets = snap.mirrored_packets;
-                binding.mirrored_bytes = snap.mirrored_bytes;
-                binding.mirror_drops_no_frame = snap.mirror_drops_no_frame;
-                binding.mirror_drops_tx_frame_reserve = snap.mirror_drops_tx_frame_reserve;
-                binding.mirror_drops_no_binding = snap.mirror_drops_no_binding;
-                binding.mirror_drops_queue_full = snap.mirror_drops_queue_full;
-                binding.mirror_drops_queue_full_same_worker =
-                    snap.mirror_drops_queue_full_same_worker;
-                binding.mirror_drops_queue_full_cross_worker =
-                    snap.mirror_drops_queue_full_cross_worker;
-                binding.post_drain_backup_bytes = snap.post_drain_backup_bytes;
-                binding.drain_sent_bytes_shaped_unconditional =
-                    snap.drain_sent_bytes_shaped_unconditional;
-                binding.post_drain_backup_cos_drops = snap.post_drain_backup_cos_drops;
-                binding.post_drain_backup_cos_drop_bytes = snap.post_drain_backup_cos_drop_bytes;
-                // #710: `snap.no_owner_binding_drops` is not copied into
-                // per-binding status — it is summed across all bindings
-                // into `ProcessStatus::cos_no_owner_binding_drops_total`
-                // at the refresh_status callsite, which is the correct
-                // operator-facing scope for this counter.
-                binding.direct_tx_packets = snap.direct_tx_packets;
-                binding.copy_tx_packets = snap.copy_tx_packets;
-                binding.in_place_tx_packets = snap.in_place_tx_packets;
-                binding.in_place_vlan_push_desc_packets = snap.in_place_vlan_push_desc_packets;
-                binding.in_place_vlan_pop_desc_packets = snap.in_place_vlan_pop_desc_packets;
-                binding.in_place_vlan_push_no_headroom_packets =
-                    snap.in_place_vlan_push_no_headroom_packets;
-                binding.in_place_l2_memmove_fallback_packets =
-                    snap.in_place_l2_memmove_fallback_packets;
-                binding.direct_tx_no_frame_fallback_packets =
-                    snap.direct_tx_no_frame_fallback_packets;
-                binding.direct_tx_build_fallback_packets = snap.direct_tx_build_fallback_packets;
-                binding.direct_tx_disallowed_fallback_packets =
-                    snap.direct_tx_disallowed_fallback_packets;
-                binding.debug_pending_fill_frames = snap.debug_pending_fill_frames;
-                binding.debug_spare_fill_frames = 0;
-                binding.debug_free_tx_frames = snap.debug_free_tx_frames;
-                binding.debug_pending_tx_prepared = snap.debug_pending_tx_prepared;
-                binding.debug_pending_tx_local = snap.debug_pending_tx_local;
-                binding.debug_outstanding_tx = snap.debug_outstanding_tx;
-                binding.tx_completion_ring_available = snap.tx_completion_ring_available;
-                binding.tx_completion_ring_available_max = snap.tx_completion_ring_available_max;
-                binding.debug_in_flight_recycles = snap.debug_in_flight_recycles;
-                // #878: per-binding capacities + in-flight gauge flow
-                // into BindingStatus so the daemon's fwdstatus
-                // Buffer% can compute UMEM and TX-ring fill ratios.
-                binding.umem_total_frames = snap.umem_total_frames;
-                binding.tx_ring_capacity = snap.tx_ring_capacity;
-                binding.umem_inflight_frames = snap.umem_inflight_frames;
-                // #802: ring-pressure counters — atomic mirrors of
-                // worker-local counters, published on the worker's
-                // per-second debug tick. `outstanding_tx` aliases
-                // `debug_outstanding_tx` for the operator-facing name.
-                binding.dbg_tx_ring_full = snap.dbg_tx_ring_full;
-                binding.dbg_sendto_enobufs = snap.dbg_sendto_enobufs;
-                // #804: split counters — bound-pending FIFO vs CoS
-                // queue admission. Pre-#804 a single `dbg_pending_overflow`
-                // was published; the wire name was removed because
-                // the semantics were ambiguous for operators.
-                binding.dbg_bound_pending_overflow = snap.dbg_bound_pending_overflow;
-                binding.dbg_cos_queue_overflow = snap.dbg_cos_queue_overflow;
-                binding.rx_fill_ring_empty_descs = snap.rx_fill_ring_empty_descs;
-                binding.outstanding_tx = snap.debug_outstanding_tx;
-                // #812: per-queue TX submit→completion latency
-                // telemetry. Materialize the fixed-cap snapshot
-                // array into a freshly-owned Vec<u64> on the wire
-                // boundary — reuses the buffer in-place to avoid
-                // allocator churn when the BindingStatus entry is
-                // refreshed on the ~1s poll cadence.
-                binding
-                    .tx_submit_latency_hist
-                    .resize(snap.tx_submit_latency_hist.len(), 0);
-                binding
-                    .tx_submit_latency_hist
-                    .copy_from_slice(&snap.tx_submit_latency_hist);
-                binding.tx_submit_latency_count = snap.tx_submit_latency_count;
-                binding.tx_submit_latency_sum_ns = snap.tx_submit_latency_sum_ns;
-                // #825: per-kick `sendto` latency telemetry mirrors
-                // the #812 submit-latency copy path above. Resize
-                // the operator-facing Vec<u64> to match the
-                // snapshot's fixed-cap array, then copy bucket
-                // counts and scalars. `tx_kick_retry_count` is the
-                // EAGAIN/EWOULDBLOCK tally (T1 ring-pushback).
-                binding
-                    .tx_kick_latency_hist
-                    .resize(snap.tx_kick_latency_hist.len(), 0);
-                binding
-                    .tx_kick_latency_hist
-                    .copy_from_slice(&snap.tx_kick_latency_hist);
-                binding.tx_kick_latency_count = snap.tx_kick_latency_count;
-                binding.tx_kick_latency_sum_ns = snap.tx_kick_latency_sum_ns;
-                binding.tx_kick_retry_count = snap.tx_kick_retry_count;
-                binding.last_heartbeat = snap.last_heartbeat;
-                binding.last_error = snap.last_error;
-                binding.ready = binding.registered
-                    && binding.bound
-                    && binding.xsk_registered
-                    && heartbeat_fresh(snap.last_heartbeat);
-            } else {
-                binding.bound = false;
-                binding.xsk_registered = false;
-                binding.xsk_bind_mode.clear();
-                binding.zero_copy = false;
-                binding.socket_fd = 0;
-                binding.socket_ifindex = 0;
-                binding.socket_queue_id = 0;
-                binding.socket_bind_flags = 0;
-                binding.rx_packets = 0;
-                binding.rx_bytes = 0;
-                binding.rx_batches = 0;
-                binding.rx_wakeups = 0;
-                binding.metadata_packets = 0;
-                binding.metadata_errors = 0;
-                binding.validated_packets = 0;
-                binding.validated_bytes = 0;
-                binding.local_delivery_packets = 0;
-                binding.forward_candidate_packets = 0;
-                binding.route_miss_packets = 0;
-                binding.neighbor_miss_packets = 0;
-                binding.discard_route_packets = 0;
-                binding.next_table_packets = 0;
-                binding.exception_packets = 0;
-                binding.config_gen_mismatches = 0;
-                binding.fib_gen_mismatches = 0;
-                binding.unsupported_packets = 0;
-                binding.flow_cache_hits = 0;
-                binding.flow_cache_misses = 0;
-                binding.flow_cache_evictions = 0;
-                binding.flow_cache_collision_evictions = 0;
-                binding.active_flow_count = 0;
-                binding.flow_cache_capacity = 0;
-                binding.v_min_throttle_hard_cap_overrides = 0;
-                binding.v_min_throttles = 0;
-                binding.session_hits = 0;
-                binding.session_misses = 0;
-                binding.session_creates = 0;
-                binding.session_expires = 0;
-                binding.session_delta_pending = 0;
-                binding.session_delta_generated = 0;
-                binding.session_delta_dropped = 0;
-                binding.session_delta_drained = 0;
-                binding.policy_denied_packets = 0;
-                binding.screen_drops = 0;
-                binding.syn_cookie_challenges = 0;
-                binding.syn_cookie_secret_unavailable = 0;
-                binding.syn_cookie_syn_ack_sent = 0;
-                binding.syn_cookie_ack_rst_sent = 0;
-                binding.syn_cookie_reply_budget_drops = 0;
-                binding.syn_cookie_ack_valid = 0;
-                binding.syn_cookie_ack_invalid = 0;
-                binding.syn_cookie_bypass = 0;
-                binding.snat_packets = 0;
-                binding.dnat_packets = 0;
-                binding.slow_path_packets = 0;
-                binding.slow_path_bytes = 0;
-                binding.slow_path_local_delivery_packets = 0;
-                binding.slow_path_missing_neighbor_packets = 0;
-                binding.slow_path_no_route_packets = 0;
-                binding.slow_path_next_table_packets = 0;
-                binding.slow_path_forward_build_packets = 0;
-                binding.slow_path_drops = 0;
-                binding.slow_path_rate_limited = 0;
-                binding.kernel_rx_dropped = 0;
-                binding.kernel_rx_invalid_descs = 0;
-                binding.tx_packets = 0;
-                binding.tx_bytes = 0;
-                binding.tx_completions = 0;
-                binding.tx_errors = 0;
-                binding.tx_shared_recycle_unknown_slot_drops = 0;
-                binding.mirrored_packets = 0;
-                binding.mirrored_bytes = 0;
-                binding.mirror_drops_no_frame = 0;
-                binding.mirror_drops_tx_frame_reserve = 0;
-                binding.mirror_drops_no_binding = 0;
-                binding.mirror_drops_queue_full = 0;
-                binding.mirror_drops_queue_full_same_worker = 0;
-                binding.mirror_drops_queue_full_cross_worker = 0;
-                binding.post_drain_backup_bytes = 0;
-                binding.drain_sent_bytes_shaped_unconditional = 0;
-                binding.post_drain_backup_cos_drops = 0;
-                binding.post_drain_backup_cos_drop_bytes = 0;
-                binding.direct_tx_packets = 0;
-                binding.copy_tx_packets = 0;
-                binding.in_place_tx_packets = 0;
-                binding.in_place_vlan_push_desc_packets = 0;
-                binding.in_place_vlan_pop_desc_packets = 0;
-                binding.in_place_vlan_push_no_headroom_packets = 0;
-                binding.in_place_l2_memmove_fallback_packets = 0;
-                binding.direct_tx_no_frame_fallback_packets = 0;
-                binding.direct_tx_build_fallback_packets = 0;
-                binding.direct_tx_disallowed_fallback_packets = 0;
-                binding.debug_pending_fill_frames = 0;
-                binding.debug_spare_fill_frames = 0;
-                binding.debug_free_tx_frames = 0;
-                binding.debug_pending_tx_prepared = 0;
-                binding.debug_pending_tx_local = 0;
-                binding.debug_outstanding_tx = 0;
-                binding.tx_completion_ring_available = 0;
-                binding.tx_completion_ring_available_max = 0;
-                binding.debug_in_flight_recycles = 0;
-                // #878: capacities + in-flight gauge zero when the
-                // binding has no live state (slot unregistered). The
-                // daemon treats zero umem_total_frames as "unknown"
-                // and falls back to the legacy Buffer% display.
-                binding.umem_total_frames = 0;
-                binding.tx_ring_capacity = 0;
-                binding.umem_inflight_frames = 0;
-                // #802: ring-pressure counters — zero when the binding
-                // has no live state (unregistered slot).
-                binding.dbg_tx_ring_full = 0;
-                binding.dbg_sendto_enobufs = 0;
-                binding.dbg_bound_pending_overflow = 0;
-                binding.dbg_cos_queue_overflow = 0;
-                binding.rx_fill_ring_empty_descs = 0;
-                binding.outstanding_tx = 0;
-                // #812: zero the submit-latency histogram when the
-                // binding has no live state (unregistered slot).
-                binding.tx_submit_latency_hist.clear();
-                binding.tx_submit_latency_count = 0;
-                binding.tx_submit_latency_sum_ns = 0;
-                // #825: zero the kick-latency histogram + retry
-                // counter when the binding has no live state.
-                binding.tx_kick_latency_hist.clear();
-                binding.tx_kick_latency_count = 0;
-                binding.tx_kick_latency_sum_ns = 0;
-                binding.tx_kick_retry_count = 0;
-                binding.last_heartbeat = None;
-                binding.last_error.clear();
-                binding.ready = false;
-            }
-        }
-        self.refresh_cos_owner_worker_map_from_binding_statuses(bindings);
-    }
 }
 
 // #710: pure-function extraction of the coordinator-level aggregation
@@ -1634,7 +802,7 @@ fn unique_interface_owner_worker_id(queues: &[crate::protocol::CoSQueueStatus]) 
     owner_worker_id
 }
 
-fn build_cos_owner_worker_by_queue(
+pub(super) fn build_cos_owner_worker_by_queue(
     forwarding: &ForwardingState,
     workers: &BTreeMap<u32, Vec<BindingPlan>>,
 ) -> BTreeMap<(i32, u8), u32> {
@@ -1665,7 +833,7 @@ fn build_worker_binding_ifindexes_from_identities(
     out
 }
 
-fn build_mirror_target_map(
+pub(super) fn build_mirror_target_map(
     identities: &BTreeMap<u32, BindingIdentity>,
     live: &BTreeMap<u32, Arc<BindingLiveState>>,
 ) -> MirrorTargetMap {
@@ -1738,7 +906,7 @@ fn build_cos_owner_worker_by_queue_with_fallback_ifindexes(
     owner_by_queue
 }
 
-fn build_cos_active_shards_by_egress_ifindex_with_fallback_ifindexes(
+pub(super) fn build_cos_active_shards_by_egress_ifindex_with_fallback_ifindexes(
     forwarding: &ForwardingState,
     preferred_worker_binding_ifindexes: &BTreeMap<u32, std::collections::BTreeSet<i32>>,
     fallback_worker_binding_ifindexes: &BTreeMap<u32, std::collections::BTreeSet<i32>>,
